Fix the test case of HessianProtocol does not pass on some machines.
